remove external DH dependency in FDI migrations
