+++ fdi/models/metadata.py
@@ -7,6 +7,7 @@ from django_countries.fields import CountryField
 from django_countries import Countries as DjCountries
 
 from fdi.models.constants import MAX_LENGTH
+from fdi.utils import datahub_country_iso_code
 
 
 class BaseMetadataModel(models.Model):
@@ -35,62 +36,6 @@ class Country(BaseMetadataModel):
     # this needs an pre_save method to set the iso code based on
     # the country name
 
-    def try_set_iso_code(self):
-        """
-        Here are the list of country where there is a mismatch between
-        Datahub data and Django Countries
-        These needs a proper fix
-
-        Aland Islands
-        BLANK
-        British Virgin Islands
-        Burkina
-        Burma
-        Cape Verde
-        Congo (Democratic Republic)
-        East Timor
-        Falkland Islands
-        Gambia, The
-        Heard Island and McDonald Island
-        Hong Kong (SAR)
-        Ivory Coast
-        Korea (North)
-        Korea (South)
-        Macao (SAR)
-        Micronesia
-        Netherlands Antilles
-        Occupied Palestinian Territories
-        Pitcairn, Henderson, Ducie and Oeno Islands
-        Reunion
-        South Georgia and South Sandwich Islands
-        St Barthelemy
-        St Helena
-        St Kitts and Nevis
-        St Lucia
-        St Martin
-        St Pierre and Miquelon
-        St Vincent
-        Sudan, South
-        Surinam
-        Svalbard and Jan Mayen Islands
-        TEST
-        United States
-        Vatican City
-        Virgin Islands (US)
-        """
-        if self.name == "United States":
-            dh_country_name = "United States of America"
-        elif self.name == "Korea (South)":
-            dh_country_name = "South Korea"
-        elif self.name == "Hong Kong (SAR)":
-            dh_country_name = "Hong Kong"
-        else:
-            dh_country_name = self.name
-
-        code = DjCountries().by_name(dh_country_name)
-        if code:
-            self.iso_code = code
-
     def save(self, **kwargs):
-        self.try_set_iso_code()
+        self.iso_code = datahub_country_iso_code(self.name)
         super().save(**kwargs)
